OSD-17209: Update backplane cli commands default log level
